+++ mtuq/grid/__init__.py
@@ -1,0 +1,8 @@
+
+from mtuq.grid.base import Grid, UnstructuredGrid
+
+from mtuq.grid.moment_tensor import FullMomentTensorGridRegular, FullMomentTensorGridRandom,\
+    DoubleCoupleGridRegular, DoubleCoupleGridRandom
+
+from mtuq.grid.force import ForceGridRegular, ForceGridRandom
+
